bq_multi_agent_app/requirements.txt
google-cloud-aiplatform[agent_engines,adk]>=1.112.0
